Replaced Npgsql with Sqlite thanks to it being easier for the program to be self-contained with the latter while also cutting down start-up times.
Each ListContactItem component fails to bind to _initialsVisibility most likely because it's not a propdp and not public.

Implemented Material Design 3 Colors by adding them as individual resource dictionaries in .csproj. Might be a good idea to use the already installed nuget package for this.

Have not bound the SelectedContactViewModel to the Contacts List yet.